test(context): move mainAgentOnly settings mapping to isolated file
Drop the flaky in-suite test that collided with leaked ModeManager mocks; cover the settings mapping in context-config-loader.test.ts with a local mock.module stub.
